backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
spacy==3.7.2
pandas==2.1.3
python-dotenv==1.0.0
email-validator==2.1.0
bleach==6.0.0
httpx==0.25.1